apps/api/requirements.txt
# DoD Procurement Document Generation System - Backend Requirements
# Consolidated requirements for FastAPI backend and AI agents

# ============================================================================
# FastAPI Backend & Web Server
# ============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2

# ============================================================================
# Database & ORM
# ============================================================================
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# ============================================================================
# Authentication & Security
# ============================================================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ============================================================================
# Data Validation & Configuration
# ============================================================================
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv>=1.0.0

# ============================================================================
# Task Queue & Caching
# ============================================================================
redis==5.0.1
celery==5.3.4

# ============================================================================
# AI & LLM Providers
# ============================================================================
anthropic>=0.18.0
openai>=1.3.0
langchain>=0.1.0
langchain-anthropic>=0.1.0

# ============================================================================
# RAG & Embeddings
# ============================================================================
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Optional: Better embeddings (choose one)
# voyageai>=0.2.0        # Voyage AI (recommended for RAG)

# Optional: GPU support (if you have NVIDIA GPU)
# faiss-gpu>=1.7.4       # Replace faiss-cpu with this

# ============================================================================
# Document Processing & Generation
# ============================================================================
PyPDF2>=3.0.0
pdfrw>=0.4
markdown>=3.4.0
pdfkit>=1.0.0
python-docx>=1.1.0
reportlab>=4.0.0

# Docling - Advanced document processing
# Provides superior PDF understanding, table extraction, and multi-format support
docling>=2.0.0

# Pin numpy to 1.x for compatibility with PyTorch/transformers
# (numpy 2.x breaks torch which was compiled against numpy 1.x)
numpy<2.0.0

# ============================================================================
# Data Processing
# ============================================================================
pandas>=2.0.0
openpyxl>=3.1.0

# ============================================================================
# Web Search & Research
# ============================================================================
tavily-python>=0.3.0
